requirements.txt
boto3==1.26.79
botocore==1.29.79
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Deprecated==1.2.13
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Limiter==3.3.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
limits==3.2.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
openpyxl==3.1.1
ordered-set==4.1.0
packaging==23.0
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.28.2
rich==13.3.1
s3transfer==0.6.0
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.15.0
zipp==3.12.0
